+++ src/modules/vehicle/Vehicle.cpp
@@ -9,45 +9,45 @@
 #include "faker-cxx/String.h"
 #include "VehicleData.h"
 
-namespace faker
+namespace faker::vehicle
 {
 
-std::string_view Vehicle::bicycle()
+std::string_view bicycle()
 {
-    return Helper::arrayElement(vehicle::bicycle_types);
+    return Helper::arrayElement(bicycle_types);
 }
 
-std::string_view Vehicle::color()
+std::string_view color()
 {
-    return Helper::arrayElement(vehicle::vehicle_colors);
+    return Helper::arrayElement(vehicle_colors);
 }
 
-std::string_view Vehicle::fuel()
+std::string_view fuel()
 {
-    return Helper::arrayElement(vehicle::fuel_types);
+    return Helper::arrayElement(fuel_types);
 }
 
-std::string_view Vehicle::manufacturer()
+std::string_view manufacturer()
 {
-    return Helper::arrayElement(vehicle::manufacturers);
+    return Helper::arrayElement(manufacturers);
 }
 
-std::string_view Vehicle::model()
+std::string_view model()
 {
-    return Helper::arrayElement(vehicle::models);
+    return Helper::arrayElement(models);
 }
 
-std::string_view Vehicle::type()
+std::string_view type()
 {
-    return Helper::arrayElement(vehicle::vehicle_types);
+    return Helper::arrayElement(vehicle_types);
 }
 
-std::string Vehicle::vehicle()
+std::string vehicleName()
 {
     return FormatHelper::format("{} {}", manufacturer(), model());
 }
 
-std::string Vehicle::vin()
+std::string vin()
 {
     std::string exclude_characters{"oiqOIQ"};
 
@@ -57,7 +57,7 @@ std::string Vehicle::vin()
                                 Number::integer(10000, 99999));
 }
 
-std::string Vehicle::vrm()
+std::string vrm()
 {
     return FormatHelper::format("{}{}{}", String::alpha(2, StringCasing::Upper), String::numeric(2, true),
                                 String::alpha(3, StringCasing::Upper));