+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to this project will be documented in this file
 * changed classes to namespaces for functions grouping, for example use `number::integer` instead of `Number::integer`, applies to all modules
 * changed std::string to std::string_view in functions where is was possible
 * changed function name from `sport` to `sportName` in sport module
+* changed function name from `vehicle` to `vehicleName` in vehicle module
 
 ### Features